lab10/back/requirements.txt
Django~=3.2.10
djangorestframework~=3.15.1
pip~=24.0
pytz~=2024.1
sqlparse~=0.4.4
asgiref~=3.8.1
setuptools~=65.5.0